requirements.txt
azure-common==1.1.26
azure-core==1.21.1
azure-keyvault==4.1.0
azure-mgmt-resource==2.2.0
azure-mgmt-storage==19.0.0
azure-identity==1.7.1
msrestazure==0.6.4
azure-mgmt-core==1.3.0